Rename module Selenium.ScrollBehaviuor ⇒ Selenium.ScrollBehavior
Well, I do not have a preference between the British or the American
spelling, but in correcting the misspelt version, I chose to match the
filename (which was spelt in the British way). But if folks prefer me to
change ot to "Behavior", I'm of course fine with that as well.